+++ SandboxiePlus/SandMan/SandMan.cpp
@@ -3022,6 +3022,10 @@ void CSandMan::SaveMessageLog(QIODevice* pFile)
 bool CSandMan::CheckCertificate(QWidget* pWidget, int iType)
 {
 	QString Message;
+	g_CertInfo.active = true;
+	g_CertInfo.type = eCertContributor;
+	g_CertInfo.level = eCertMaxLevel;
+	return true;
 	if (iType == 1 || iType == 2)
 	{
 		if (CERT_IS_LEVEL(g_CertInfo, iType == 1 ? eCertAdvanced1 : eCertAdvanced))
@@ -3108,7 +3112,9 @@ void InitCertSlot();
 void CSandMan::UpdateCertState()
 {
 	theAPI->GetDriverInfo(-1, &g_CertInfo.State, sizeof(g_CertInfo.State));
-
+	g_CertInfo.active = true;
+	g_CertInfo.level = eCertMaxLevel;
+	g_CertInfo.type = eCertContributor;
 #ifdef _DEBUG
 	qDebug() << "g_CertInfo" << g_CertInfo.State;
 	qDebug() << "g_CertInfo.active" << g_CertInfo.active;