+++ CHANGELOG.md
@@ -10,7 +10,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - added alias for a sandbox [#4112](https://github.com/sandboxie-plus/Sandboxie/issues/4112)
-
+- added UI options for "ForceRestartAll" and "UseCreateToken" in OptionWindow
+- added an optional context menu option to make folder/file forced quickly
+  - Note: You can also use "Sandman.exe /add_force program_path" to do it
+- added a feature which could modify Windows Product Id in the registry to a rand value