+++ CHANGELOG.md
@@ -13,7 +13,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
-
 ## [1.14.5 / 5.69.5] - 2024-07-23
 
 ### Added
@@ -53,6 +52,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+
 ## [1.14.3 / 5.69.3] - 2024-07-01
 
 ### Changed
@@ -84,6 +84,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - added RPC Port message filter mechanism to block unsafe RDP calls via the driver [#3930](https://github.com/sandboxie-plus/Sandboxie/issues/3930)
   - Usage: "RpcPortFilter=Port,ID,Label" label is optional
 - added "Job Object" Options page to collect all job object-related options
+
 
 ### Changed
 - extended "Temp Template" to make it could delete local template section